requirements.txt
fastapi
uvicorn
supabase
apscheduler
httpx
pandas
scikit-learn
numpy
python-dotenv
joblib
scipy